services/core/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.1
websockets>=13.0
python-dotenv==1.0.0
httpx==0.27.0
supabase==2.3.0
google-cloud-speech==2.25.1
google-cloud-translate==3.15.0
google-cloud-texttospeech==2.16.0
livekit-api==0.5.1
python-multipart==0.0.9
pydantic-settings
PyJWT
email-validator